Move display initialization out of init
I2C failures would raise due to the use of `with`, but the error would
due to `init/1` returning the wrong thing. This move I2C initialization
to a `handle_continue/2` and checks error codes so that code that raises
will give a stack trace with what failed and not get into a tight init
exception loop.